Fix apt and swap check

+++ Vagrant.bootstrap.sh
@@ -3,7 +3,7 @@
 ### Aprovisionamiento de software ###
 
 # Actualizo los paquetes de la maquina virtual
-sudo apt-get updates
+sudo apt-get update
 
 # Instalo un servidor web
 sudo apt-get install -y apache2 
@@ -11,17 +11,17 @@ sudo apt-get install -y apache2
 ### Configuración del entorno ###
 
 ##Genero una partición swap. Previene errores de falta de memoria
-sudo mkdir /swapdir
-cd /swapdir
-sudo dd if=/dev/zero of=/swapdir/swapfile bs=1024 count=2000000
-sudo mkswap -f  /swapdir/swapfile
-sudo chmod 600 /swapdir/swapfile
-sudo swapon swapfile
-echo "/swapdir/swapfile       none    swap    sw      0       0" | sudo tee -a /etc/fstab /etc/fstab
-sudo sysctl vm.swappiness=10
-echo vm.swappiness = 10 | sudo tee -a /etc/sysctl.conf
-
-
+if [ ! -f "/swapdir/swapfile" ]; then
+	sudo mkdir /swapdir
+	cd /swapdir
+	sudo dd if=/dev/zero of=/swapdir/swapfile bs=1024 count=2000000
+	sudo mkswap -f  /swapdir/swapfile
+	sudo chmod 600 /swapdir/swapfile
+	sudo swapon swapfile
+	echo "/swapdir/swapfile       none    swap    sw      0       0" | sudo tee -a /etc/fstab /etc/fstab
+	sudo sysctl vm.swappiness=10
+	echo vm.swappiness = 10 | sudo tee -a /etc/sysctl.conf
+fi
 
 # ruta raíz del servidor web
 APACHE_ROOT="/var/www";
